Start Creating new Main Menu
Currently under Adam Scripts and Adam Scenes, I am working to revamp the Main Menu so that it not only suits our needs but we can also adapt it to new changes easily and have a more organized system. Right now the only button that works is the quit button. Next push will be the Multiplayer Button creating a lobby menu