Add MusicManager
Game Manager remake now will check which state of the game it is in in terms of MainMenu, Level. The music manager only takes care of the music depending on which gamestate the players are in.

Also, bouncing obstacles and balloons do not use the old audiomanager anymore and have their own dedicated audiosource, at least the ones in the sample scene. All objects from this point forward should have their own audiosource if they are suppose to make sounds.